modify: firebase api 리팩토링 - 쿼리 빌더 함수를 util 로 분리
